# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Resource-Service is service for storing kubernetes resources in Mongo DB.",
    "title": "Resource-Service",
    "version": "0.0.1"
  },
  "paths": {
...
    "/namespaces/{namespace}/ingresses": {
      "get": {
        "tags": [
          "Ingress"
        ],
        "summary": "Get ingresses list.",
        "operationId": "GetIngressesListHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ingresses list",
            "schema": {
              "$ref": "#/definitions/IngressesResponse"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Ingress"
        ],
        "summary": "Create ingress.",
        "operationId": "CreateIngressHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Ingress"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ingress created",
            "schema": {
              "$ref": "#/definitions/ResourceIngress"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "Ingress"
        ],
        "summary": "Delete all ingresses.",
        "operationId": "DeleteAllIngressesHandler",
        "parameters": [
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "all ingresses in namespace deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "private"
      }
    },
    "/namespaces/{namespace}/ingresses/{ingress}": {
      "get": {
        "tags": [
          "Ingress"
        ],
        "summary": "Get ingresses list.",
        "operationId": "GetIngressHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ingress",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ingresses",
            "schema": {
              "$ref": "#/definitions/ResourceIngress"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "Ingress"
        ],
        "summary": "Update ingress.",
        "operationId": "UpdateIngressHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ingress",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Ingress"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "ingress updated",
            "schema": {
              "$ref": "#/definitions/ResourceIngress"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      },
      "delete": {
        "tags": [
          "Ingress"
        ],
        "summary": "Delete ingress.",
        "operationId": "DeleteIngressHandler",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "X-User-Namespace",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ingress",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "ingress deleted"
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/errOAIGen"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
...
  },
  "definitions": {
...
    "Ingress": {
      "description": "Ingress -- model for ingress",
      "type": "object",
      "required": [
        "name",
        "rules"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "rules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Rule"
          },
          "x-go-name": "Rules"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "IngressesResponse": {
      "description": "IngressesResponse -- ingresses response",
      "type": "object",
      "properties": {
        "ingresses": {
          "$ref": "#/definitions/ListIngress"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/ingress"
    },
...
    "ListIngress": {
      "description": "ListIngress -- ingresses list",
      "type": "array",
      "items": {
        "$ref": "#/definitions/ResourceIngress"
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/ingress"
    },
...
    "Path": {
      "description": "Path -- ingress path",
      "type": "object",
      "required": [
        "path",
        "service_name",
        "service_port"
      ],
      "properties": {
        "path": {
          "type": "string",
          "x-go-name": "Path"
        },
        "service_name": {
          "type": "string",
          "x-go-name": "ServiceName"
        },
        "service_port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ServicePort"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ResourceIngress": {
      "description": "ResourceIngress --  model for ingress for resource-service db",
      "type": "object",
      "required": [
        "name",
        "rules"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "deleted": {
          "type": "boolean",
          "x-go-name": "Deleted"
        },
        "deleted_at": {
          "description": "delete date in RFC3339 format",
          "type": "string",
          "x-go-name": "DeletedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        },
        "namespaceid": {
          "type": "string",
          "x-go-name": "NamespaceID"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "rules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Rule"
          },
          "x-go-name": "Rules"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/pkg/models/ingress"
    },
...
    "Rule": {
      "description": "Rule -- ingress rule",
      "type": "object",
      "required": [
        "host",
        "path"
      ],
      "properties": {
        "host": {
          "type": "string",
          "x-go-name": "Host"
        },
        "path": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Path"
          },
          "x-go-name": "Path"
        },
        "tls_secret": {
          "type": "string",
          "x-go-name": "TLSSecret"
        }
      },
      "x-go-package": "git.containerum.net/ch/resource-service/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "errOAIGen": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "description": "Fields -- represents additional key-value fields of error",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "id": {
          "$ref": "#/definitions/errOaiGenId"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "github.com/containerum/cherry"
    },
    "errOaiGenId": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "description": "ErrKind -- represents kind of error",
          "type": "integer",
          "format": "uint64",
          "x-go-package": "github.com/containerum/cherry"
        },
        "sid": {
          "description": "ErrSID -- represents service ID of error",
          "type": "string",
          "x-go-package": "github.com/containerum/cherry"
        }
      },
      "x-go-gen-location": "models",
      "x-go-package": "github.com/containerum/cherry"
    }
  },
...
}